Allow to not sync empty translations
Usually empty messages should not be written to the filesystem otherwise the symfony translator fallback cannot do it's job.
For not breaking current behaviour, it's possible to configure this option and the default will be true (write empty messages to filesystem).
In a major update of this bundle, the default could be set to false.